// src/utility/enums.ts
// Enums for Transactions and Categories

export enum TransactionCategory {
  Income = "Income",
  Expense = "Expense"
}

export enum IncomeCategory {
  Business = 'Business',
  Investments = 'Investments',
  ExtraIncome = 'Extra income',
  Deposits = 'Deposits',
  Lottery = 'Lottery',
  Gifts = 'Gifts',
  Salary = 'Salary',
  Savings = 'Savings',
  RentalIncome = 'Rental income'
}

export enum ExpenseCategory {
  Bills = 'Bills',
  Car = 'Car',
  Clothes = 'Clothes',
  Travel = 'Travel',
  Food = 'Food',
  Shopping = 'Shopping',
  House = 'House',
  Entertainment = 'Entertainment',
  Phone = 'Phone',
  Pets = 'Pets',
  Other = 'Other'
}